fix: follow suggestions
